feat/SEOULDATA-65 rename: 도메인 구조로 클래스 이동 및 분리
