--- fest/src/main/java/com/seouldata/fest/domain/fest/service/ReviewService.java
@@ -1,9 +1,0 @@
-package com.seouldata.fest.domain.fest.service;
-
-import com.seouldata.fest.domain.fest.dto.resquest.AddReviewReq;
-
-public interface ReviewService {
-
-    Long addReview(Long memSeq, AddReviewReq addReviewReq);
-
-}
