+++ fest/src/main/java/com/seouldata/fest/domain/review/entity/Review.java
@@ -1,6 +1,8 @@
-package com.seouldata.fest.domain.fest.entity;
+package com.seouldata.fest.domain.review.entity;
 
+import com.seouldata.fest.domain.fest.entity.Fest;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -9,27 +11,32 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
 
 @Entity
-@Table(name = "tag")
+@Table(name = "review")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Tag {
+public class Review {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "tag_seq")
-    private Long tagSeq;
+    @Column(name = "review_seq")
+    private Long reviewSeq;
 
     @Column(name = "mem_seq")
     @NotNull
     private Long memSeq;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "review_seq", nullable = false)
-    private Review review;
+    @JoinColumn(name = "fest_seq", nullable = false)
+    private Fest fest;
 
-    @Column(name = "tag_no", columnDefinition = "tinyint(4)")
+    @Column(columnDefinition = "smallint(6)")
     @NotNull
-    private int tagNo;
+    @Max(5)
+    private int point;
+
+    @Column(length = 150)
+    @NotNull
+    private String content;
 
     @Column(name = "is_deleted", columnDefinition = "tinyint(1)")
     @NotNull
@@ -37,10 +44,11 @@ public class Tag {
     private boolean isDeleted;
 
     @Builder
-    public Tag(Long memSeq, Review review, int tagNo, boolean isDeleted) {
+    public Review(Long memSeq, Fest fest, int point, String content, boolean isDeleted) {
         this.memSeq = memSeq;
-        this.review = review;
-        this.tagNo = tagNo;
+        this.fest = fest;
+        this.point = point;
+        this.content = content;
         this.isDeleted = isDeleted;
     }
