+++ fest/src/test/java/com/seouldata/fest/domain/fest/repository/FestRepositoryTest.java
@@ -1,7 +1,6 @@
 package com.seouldata.fest.domain.fest.repository;
 
-import com.seouldata.fest.domain.fest.entity.*;
-import com.seouldata.fest.domain.fest.repository.*;
+import com.seouldata.fest.domain.fest.entity.Fest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -15,19 +14,6 @@ public class FestRepositoryTest {
 
     @Autowired
     private FestRepository festRepository;
-
-    @Autowired
-    private ReviewRepository reviewRepository;
-
-    @Autowired
-    private TagRepository tagRepository;
-
-    @Autowired
-    private ImageRepository imageRepository;
-
-    @Autowired
-    private HeartRepository heartRepository;
-
 
     @Test
     public void InsertFestDummy() {
@@ -49,56 +35,6 @@ public class FestRepositoryTest {
                 .build();
 
         festRepository.save(fest);
-    }
-
-    @Test
-    public void InsertReviewDummy() {
-
-        Review review = Review.builder()
-                .memSeq(1L)
-                .fest(festRepository.findById(1L).get())
-                .point(5)
-                .content("정말 최고입니다")
-                .isDeleted(false)
-                .build();
-
-        reviewRepository.save(review);
-    }
-
-    @Test
-    public void InsertTagDummy() {
-
-        Tag tag = Tag.builder()
-                .memSeq(1L)
-                .review(reviewRepository.findById(1L).get())
-                .tagNo(1)
-                .isDeleted(false)
-                .build();
-
-        tagRepository.save(tag);
-    }
-
-    @Test
-    public void InsertImageDummy() {
-
-        Image image = Image.builder()
-                .review(reviewRepository.findById(1L).get())
-                .imgUrl("test_img")
-                .isDeleted(false)
-                .build();
-
-        imageRepository.save(image);
-    }
-
-    @Test
-    public void InsertHeartDummy() {
-
-        Heart heart = Heart.builder()
-                .memSeq(1L)
-                .fest(festRepository.findById(1L).get())
-                .build();
-
-        heartRepository.save(heart);
     }
 
 }
